Universal URL
- /cashpoint/<cashpointId>/...
- Merged LiveOrderController with CashpointController
